labs/lab01-langchain-basics/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0